Update simplified_script.py
Added a network test using 'socket' which checks if the computer can resolve the SDSS hostname into an IP address, and if so it will go on with the request unless there is some server error. Also added a random number generator for the object index to generate a random galaxy with each run.